feat: Location 페이지에 Suspense 경계 추가 (#173)
