+++ app/(main)/location/page.tsx
@@ -1,97 +1,13 @@
-'use client'
+import { Suspense } from 'react'
 
-import { useRouter, useSearchParams } from 'next/navigation'
-import { ChangeEventHandler, useEffect, useRef, useState } from 'react'
-
-import MapSection from '@/app/(main)/location/components/MapSection'
-import VehicleStatusPanel from '@/app/(main)/location/components/VehicleStatusPanel'
-import SearchInput from '@/components/common/Input/SearchInput'
-import Modal from '@/components/common/Modal'
-import { ModalMessageType } from '@/components/common/Modal/types'
-import { MAP_CONFIG } from '@/constants/map'
-import { useMapStatus } from '@/hooks/useMapStatus'
-import { useVehicleDisclosure } from '@/hooks/useVehicleDisclosure'
-import { useVehicleLocationSearch } from '@/hooks/useVehicleLocationSearch'
-import { vehicleService } from '@/lib/apis'
-import { removeSpaces, trimValue } from '@/lib/utils/string'
-import { cleanUrlParams } from '@/lib/utils/url'
-import { useVehicleVisibleStore } from '@/stores/useVehicleVisibleStore'
-import { VehicleDetail, VehicleLocation } from '@/types/vehicle'
-
-import * as styles from './styles.css'
+import Location from './Location'
 
 const LocationPage = () => {
-    const [vehicleDetail, setVehicleDetail] = useState<VehicleDetail>()
-
-    const mapRef = useRef<kakao.maps.Map>(null)
-
-    const { controlMapStatus } = useMapStatus(mapRef.current)
-    const { showSearchedVehicle, hideSelectedVehicle } = useVehicleDisclosure()
-    const { inputValue, setInputValue } = useVehicleVisibleStore()
-    const { vehicleInfo, isModalOpen, message, closeModal, searchVehicleWithNumber } =
-        useVehicleLocationSearch(inputValue)
-
-    const router = useRouter()
-    const searchParams = useSearchParams()
-
-    useEffect(() => {
-        return () => cleanUrlParams()
-    }, [])
-
-    const handleInputSubmit = async () => {
-        const vehicleInfo = await searchVehicleWithNumber()
-
-        if (!vehicleInfo) return
-
-        const { vehicleId } = vehicleInfo as VehicleLocation
-        const vehicleDetail = await vehicleService.getVehicleDetail(vehicleId)
-
-        hideSelectedVehicle()
-
-        const params = new URLSearchParams(searchParams)
-        params.set('vehicleNumber', removeSpaces(trimValue(inputValue)))
-        router.replace(`/location?${params.toString()}`)
-
-        setVehicleDetail(vehicleDetail)
-        showSearchedVehicle()
-
-        controlMapStatus(
-            {
-                lat: vehicleInfo.coordinate.lat,
-                lng: vehicleInfo.coordinate.lng,
-            },
-            MAP_CONFIG.SEARCH_VEHICLE.ZOOM_INCREMENT,
-        )
-    }
-
-    const handleInputChange: ChangeEventHandler<HTMLInputElement> = (event) => {
-        setInputValue(event.target.value)
-    }
-
     return (
-        <div className={styles.container}>
-            <MapSection
-                mapRef={mapRef}
-                vehicleInfo={vehicleInfo as VehicleLocation}
-                searchedDetail={vehicleDetail as VehicleDetail}
-            />
-            <div className={styles.searchInputWrapper}>
-                <SearchInput
-                    icon='/icons/search-icon.svg'
-                    value={inputValue}
-                    onChange={handleInputChange}
-                    onSubmit={handleInputSubmit}
-                />
-            </div>
-            <VehicleStatusPanel />
-
-            <Modal
-                isOpen={isModalOpen}
-                message={message as ModalMessageType}
-                variant={{ variant: 'alert', confirmButton: '확인' }}
-                onClose={closeModal}
-            />
-        </div>
+        // TODO: Suspense fallback 변경
+        <Suspense fallback={<div>서스펜스 불러오는 중!</div>}>
+            <Location />
+        </Suspense>
     )
 }
